halcyon-full/core/enrichment/pyproject.toml
[project]
name = "halcyon-enrichment"
version = "0.1.0"
description = "HALCYON Foundry — Enrichment Service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "asyncpg>=0.29.0",
  "pyjwt>=2.8.0",
  "prometheus-client>=0.20.0",
  "opentelemetry-api>=1.24.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-instrumentation-fastapi>=0.45b0",
  "opentelemetry-instrumentation-httpx>=0.45b0",
  "opentelemetry-exporter-otlp-proto-http>=1.24.0",
]
